- GUI: XQuery Editor bug fixes - numerous checkstyle/findbugs tests and fixes - XQuery/XPath tokens converted to Strings to simplify support of diff. encodings

+++ src/org/basex/util/Token.java
@@ -208,7 +208,7 @@ public final class Token {
    * @param v first character byte
    * @return character length
    */
-  public static int cl(final int v) {
+  public static int cl(final byte v) {
     return CHLEN[(v & 0xFF) >> 4];
   }
